+++ src/commands/hardis/project/deploy/sources/metadata.ts
@@ -5,10 +5,9 @@ import { Messages } from '@salesforce/core';
 import { AnyJson } from '@salesforce/ts-types';
 import * as c from 'chalk';
 import * as fs from 'fs-extra';
-import * as os from 'os';
-import * as path from 'path';
 import { MetadataUtils } from '../../../../../common/metadata-utils';
-import { execCommand, uxLog } from '../../../../../common/utils';
+import { createTempDir, execCommand, uxLog } from '../../../../../common/utils';
+import { deployDestructiveChanges, deployMetadatas } from '../../../../../common/utils/deployUtils';
 import { getConfig } from '../../../../../config';
 
 // Initialize Messages with the current plugin directory
@@ -93,22 +92,6 @@ export default class DxSources extends SfdxCommand {
     const destructiveProcessed = false;
     let deployProcessed = false;
 
-    // Deploy destructive changes
-    const packageDeletedXmlFile =
-      destructivePackageXml ||
-        process.env.PACKAGE_XML_TO_DELETE ||
-        this.configInfo.packageXmlToDelete ||
-        fs.existsSync('./manifest/destructiveChanges.xml')
-        ? './manifest/destructiveChanges.xml'
-        : fs.existsSync('./destructiveChanges.xml')
-          ? './destructiveChanges.xml'
-          : './config/destructiveChanges.xml';
-    if (fs.existsSync(packageDeletedXmlFile)) {
-      await MetadataUtils.deployDestructiveChanges(packageDeletedXmlFile, { debug: debugMode, check }, this);
-    } else {
-      uxLog(this, 'No destructivePackage.Xml found so no destructive deployment has been performed');
-    }
-
     // Deploy sources
     const packageXmlFile =
       packageXml ||
@@ -123,7 +106,7 @@ export default class DxSources extends SfdxCommand {
       let deployDir = '.';
       // Filter if necessary
       if (filter) {
-        const tmpDir = path.join(os.tmpdir(), 'sfdx-hardis-deploy-') + Math.random().toString(36).slice(-5);
+        const tmpDir = await createTempDir();
         const filterCommand = 'sfdx essentials:metadata:filter-from-packagexml' +
           ` -i ${deployDir}` +
           ` -p ${packageXmlFile}` +
@@ -136,7 +119,7 @@ export default class DxSources extends SfdxCommand {
         });
       }
       // Perform deployment
-      const deployRes = await MetadataUtils.deployMetadatas({
+      const deployRes = await deployMetadatas({
         deployDir,
         testlevel,
         check,
@@ -154,6 +137,22 @@ export default class DxSources extends SfdxCommand {
       }
     } else {
       uxLog(this, 'No package.xml found so no deployment has been performed');
+    }
+
+    // Deploy destructive changes
+    const packageDeletedXmlFile =
+    destructivePackageXml ||
+      process.env.PACKAGE_XML_TO_DELETE ||
+      this.configInfo.packageXmlToDelete ||
+      fs.existsSync('./manifest/destructiveChanges.xml')
+      ? './manifest/destructiveChanges.xml'
+      : fs.existsSync('./destructiveChanges.xml')
+        ? './destructiveChanges.xml'
+        : './config/destructiveChanges.xml';
+    if (fs.existsSync(packageDeletedXmlFile)) {
+      await deployDestructiveChanges(packageDeletedXmlFile, { debug: debugMode, check }, this);
+    } else {
+      uxLog(this, 'No destructivePackage.Xml found so no destructive deployment has been performed');
     }
 
     return {